--HG--
extra : convert_revision : svn%3A0fb12191-db4b-0410-8866-e78deaee71e5/trunk%403800

+++ mpqc/src/lib/util/group/memamsg.cc
@@ -187,7 +187,7 @@ ActiveMsgMemoryGrp::set_localsize(size_t localsize)
   if (debug_) {
       ExEnv::out0() << "ActiveMsgMemoryGrp::set_localsize done: offsets:";
       for (int i=0; i<=n(); i++) {
-          ExEnv::out0() << " " << offset(i);
+          ExEnv::out0() << " " << double(offset(i));
         }
       ExEnv::out0() << endl;
     }
